arm64: dts: qcom: msm8916-samsung-fortuna: Add initial device trees
Samsung Galaxy Grand Prime is a phone based on MSM8916. It is similar to
the other Samsung devices based on MSM8916 with only a few minor
differences.

This initial commit adds support for:
 - fortuna3g (SM-G530H)
 - gprimeltecan (SM-G530W)
 - grandprimelte (SM-G530FZ)

The device trees contain initial support with:
 - GPIO keys
 - Regulator haptic
 - SDHCI (internal and external storage)
 - USB Device Mode
 - UART (on USB connector via the SM5502 MUIC)
 - WCNSS (WiFi/BT)
 - Regulators
 - Sound and modem

There are different variants of Grand Prime, with some differences
in accelerometer, NFC and panel.

The common parts are shared in msm8916-samsung-fortuna-common.dtsi
to reduce duplication.

Signed-off-by: Walter Broemeling <wallebroem@gmail.com>
Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
[Joe: Add buttons and WCNSS]
Signed-off-by: Joe Mason <buddyjojo06@outlook.com>
[Siddharth: Add fortuna3g]
Signed-off-by: Siddharth Manthan <siddharth.manthan@gmail.com>
[Lin: Refactor with common dtsi. Add grandprimelte.]
Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>