drm/panel: Add driver for Samsung AMS497EE01 panel
This patch adds support for Samsung's AMOLED AMS497EE01 panel with an
EA8061V controller. The driver is mostly based on the driver for
s6e3fa2 and s6e63m0, with data taken from the downstream android
kernel for the device with codename a500f.

The panel is used in the msm8916 Samsung Galaxy A5U (EUR) devices, for
which there is already partial support in the mainline kernel.

Signed-off-by: Henrik Grimler <henrik@grimler.se>