arm64: dts: qcom: msm8916-samsung-rossa: Add initial device tree
Samsung Galaxy Core Prime LTE is a phone based on MSM8916. It is similar
to the other Samsung devices based on MSM8916 with only a few minor
differences.

Core Prime LTE and Grand Prime are similar, with some differences in MUIC,
panel and touchscreen. The common parts are shared in
msm8916-samsung-fortuna-common.dtsi to reduce duplication.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>